requirements.txt
pandas
matplotlib
seaborn
google-generativeai
openpyxl
python-dotenv
fastapi
uvicorn
python-multipart
aiofiles
matplotlib
seaborn
google-generativeai
openpyxl
python-dotenv
fastapi
uvicorn
python-multipart
aiofiles
